lift filter-clause building into ReferenceESRepository
