+++ app/domain/references/repository.py
@@ -3,11 +3,12 @@
 import datetime
 from abc import ABC
 from collections.abc import Sequence
-from typing import Literal
+from typing import ClassVar, Literal
 from uuid import UUID
 
 from elasticsearch import AsyncElasticsearch
 from elasticsearch.dsl import AsyncSearch, Q
+from elasticsearch.dsl.query import Prefix, Query, Range, Term, Terms
 from opentelemetry import trace
 from sqlalchemy import (
     CompoundSelect,
@@ -30,13 +31,18 @@ from app.domain.references.models.es import (
     RobotAutomationPercolationDocument,
 )
 from app.domain.references.models.models import (
+    AnnotationFilter,
     CandidateCanonicalSearchFields,
     DuplicateDetermination,
+    FacetType,
     GenericExternalIdentifier,
+    LinkedDataConceptFilter,
     PendingEnhancementStatus,
+    PublicationYearRange,
     ReferenceSearchProjection,
     ReferenceWithChangeset,
     RobotAutomationPercolationResult,
+    SearchQuery,
 )
 from app.domain.references.models.models import (
     Enhancement as DomainEnhancement,
@@ -88,7 +94,7 @@ from app.domain.references.models.sql import (
 )
 from app.domain.references.models.sql import SearchExport as SQLSearchExport
 from app.persistence.blob.models import BlobStorageFile
-from app.persistence.es.persistence import ESScoreResult
+from app.persistence.es.persistence import ESFacetBucket, ESScoreResult, ESSearchResult
 from app.persistence.es.repository import GenericAsyncESRepository
 from app.persistence.generics import GenericPersistenceType
 from app.persistence.repository import GenericAsyncRepository
@@ -245,6 +251,17 @@ class ReferenceESRepository(
 ):
     """Concrete implementation of a repository for references using Elasticsearch."""
 
+    #: Fields the user-supplied query string searches against by default.
+    default_search_fields: ClassVar[tuple[str, ...]] = (
+        "title",
+        "abstract",
+    )
+
+    #: Mapping from a facet type to the ES field its counts are aggregated on.
+    _FACET_FIELDS: ClassVar[dict[FacetType, str]] = {
+        FacetType.CONCEPTS: "linked_data_concepts",
+    }
+
     def __init__(self, client: AsyncElasticsearch) -> None:
         """Initialize the repository with the Elasticsearch client."""
         super().__init__(
@@ -252,6 +269,110 @@ class ReferenceESRepository(
             ReferenceSearchProjection,
             ReferenceDocument,
         )
+
+    def _build_publication_year_clause(
+        self,
+        publication_year_range: PublicationYearRange,
+    ) -> Query | None:
+        """Range clause on ``publication_year``; ``None`` if both bounds are unset."""
+        bounds: dict[str, int] = {}
+        if publication_year_range.start is not None:
+            bounds["gte"] = publication_year_range.start
+        if publication_year_range.end is not None:
+            bounds["lte"] = publication_year_range.end
+        if not bounds:
+            return None
+        return Range(publication_year=bounds)
+
+    def _build_linked_data_concept_clause(
+        self,
+        concept_filter: LinkedDataConceptFilter,
+    ) -> Query:
+        """Terms clause matching any of the listed concept URIs (OR semantics)."""
+        return Terms(linked_data_concepts=concept_filter.concept_uris)
+
+    def _build_annotation_clause(self, annotation: AnnotationFilter) -> Query:
+        """
+        Build a structured DSL clause for an annotation filter.
+
+        Three cases mirror the original Lucene builder:
+
+        - ``score`` set: range on the dynamic ``<scheme>[_<label>]`` numeric field,
+          with ``:`` in the scheme replaced by ``_`` (e.g. ``inclusion_destiny``).
+        - scheme only, no score: prefix match on the ``annotations`` keyword field,
+          matching any ``<scheme>/...`` annotation.
+        - scheme + label: exact term match on ``annotations``.
+        """
+        if annotation.score is not None:
+            field = annotation.scheme.replace(":", "_")
+            if annotation.label:
+                field += f"_{annotation.label}"
+            return Range(**{field: {"gte": annotation.score}})
+        if not annotation.label:
+            return Prefix(annotations=f"{annotation.scheme}/")
+        return Term(annotations=f"{annotation.scheme}/{annotation.label}")
+
+    def _build_filter_clauses(self, query: SearchQuery) -> list[Query]:
+        """Translate a SearchQuery's structured filters into bool.filter clauses."""
+        clauses: list[Query] = []
+        if query.publication_year_range and (
+            clause := self._build_publication_year_clause(query.publication_year_range)
+        ):
+            clauses.append(clause)
+        clauses.extend(
+            self._build_annotation_clause(annotation)
+            for annotation in query.annotation_filters
+        )
+        clauses.extend(
+            self._build_linked_data_concept_clause(concept_filter)
+            for concept_filter in query.linked_data_concept_filters
+        )
+        return clauses
+
+    @trace_repository_method(tracer)
+    async def search(
+        self,
+        query: SearchQuery,
+        page: int = 1,
+        page_size: int = 20,
+        sort: list[str] | None = None,
+    ) -> ESSearchResult:
+        """Search references matching ``query``; structured filters AND with q."""
+        return await self.search_with_query_string(
+            query.query_string,
+            fields=self.default_search_fields,
+            page=page,
+            page_size=page_size,
+            sort=sort,
+            filter_clauses=self._build_filter_clauses(query),
+            parse_document=False,
+        )
+
+    @trace_repository_method(tracer)
+    async def aggregate_facets(
+        self,
+        query: SearchQuery,
+        facets: Sequence[FacetType],
+        *,
+        max_buckets: int,
+    ) -> dict[FacetType, list[ESFacetBucket]]:
+        """
+        Count occurrences per facet over references matching ``query``.
+
+        Naive: counts are scoped by the full query, so filters within a facet
+        contribute to that facet's own counts. See destiny-repository#703.
+        """
+        facet_to_field = {facet: self._FACET_FIELDS[facet] for facet in facets}
+        buckets_by_field = await self.aggregate_terms(
+            query.query_string,
+            aggregate_on=list(facet_to_field.values()),
+            query_fields=self.default_search_fields,
+            filter_clauses=self._build_filter_clauses(query),
+            max_buckets=max_buckets,
+        )
+        return {
+            facet: buckets_by_field[field] for facet, field in facet_to_field.items()
+        }
 
     @staticmethod
     def _build_author_dis_max_query(
